requirements.txt
pandas
django==2.2.23
django-filter==2.4.0
django-tables2==2.4.0